+++ src/app/components/sheets/AlertList.tsx
@@ -1,11 +1,11 @@
-import { Alert } from '@data/datatypes';
 import { useAlerts } from '@data/queries/app';
+import useAppStore from '@data/state/app_state';
+import { Alert } from '@data/types';
 import { appLogger } from '@data/utils/logger';
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { BottomSheetFlatList } from '@gorhom/bottom-sheet';
 import React, { memo } from 'react';
 import { Text, TouchableOpacity, View } from 'react-native';
-import useAppStore from 'src/data/app_state';
 import { Sheets, useSheetController } from '../providers/sheet-controller';
 import SheetHeader from '../ui/SheetHeader';
 import BaseSheet, { SheetProps } from './BaseSheet';
