+++ src/app/components/ui/TripPlanCell.tsx
@@ -1,9 +1,9 @@
-import { IOptionDetail } from '@data/utils/interfaces';
+import { IOptionDetail } from '@data/typecheck/aggie_spirit';
 import { Ionicons } from '@expo/vector-icons';
 import React from 'react';
 import { Keyboard, Text, TouchableOpacity, View } from 'react-native';
 
-import useAppStore from 'src/data/app_state';
+import useAppStore from '@data/state/app_state';
 import { Sheets, useSheetController } from '../providers/sheet-controller';
 
 interface TripPlanCellProps {
