+++ src/app/index.tsx
@@ -1,3 +1,4 @@
+import useAppStore from '@data/state/app_state';
 import {
   BottomSheetModal,
   BottomSheetModalProvider,
@@ -6,7 +7,6 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useEffect, useRef, useState } from 'react';
 import { Appearance, BackHandler, StatusBar, View } from 'react-native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import useAppStore from 'src/data/app_state';
 import RouteMap from './components/map/MapView';
 import {
   SheetControllerContext,
