+++ src/data/typecheck/aggie_spirit.ts
